refactor: 전역 타입 디렉토리 구조 개선 [TM-34]
