+++ src/entities/main/main.types.ts
