+++ src/shared/types/hooks.types.ts
@@ -1,0 +1,6 @@
+interface IBulleetPointConverter {
+  e: React.KeyboardEvent<HTMLTextAreaElement>;
+  setInputValue: React.Dispatch<React.SetStateAction<string>>;
+}
+
+export type { IBulleetPointConverter };
