--- src/shared/types/modal.types.ts
@@ -1,9 +1,0 @@
-import { ReactNode } from 'react';
-
-interface IModal {
-  children: ReactNode;
-  isOpen: boolean;
-  toggle: () => void;
-}
-
-export type { IModal };
