+++ src/shared/types/utils.types.ts
@@ -1,0 +1,6 @@
+interface IHandleInputChangeProps {
+  e: React.ChangeEvent<HTMLTextAreaElement>;
+  setInputValue: React.Dispatch<React.SetStateAction<string>>;
+}
+
+export type { IHandleInputChangeProps };
